+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/provider-utils
 
+## 5.0.0-canary.33
+
+### Patch Changes
+
+- 2427d88: feat(ai): change Tool.sensitiveContext to telemetry.includeToolsContext and make it opt-in
+
 ## 5.0.0-canary.32
 
 ### Major Changes